backend/requirements.txt
Flask==2.1.3
flask-cors==3.0.10
Pillow==9.3.0
pydub==0.25.1
PyPDF2==3.0.1
python-docx==0.8.11
pdf2image==1.16.3
fpdf==1.7.2
Werkzeug==2.0.3
